Fix Color issues, fix frame label as clips option

+++ Editor/Flash/Records/CXFormWithAlpha.cs
@@ -2,14 +2,14 @@ namespace CWAEmu.OFUCU.Flash.Records {
     public class CXFormWithAlpha {
         private uint nBits;
 
-        public byte RMult { get; private set; } = 1;
-        public byte GMult { get; private set; } = 1;
-        public byte BMult { get; private set; } = 1;
-        public byte AMult { get; private set; } = 1;
-        public byte RAdd { get; private set; }
-        public byte GAdd { get; private set; }
-        public byte BAdd { get; private set; }
-        public byte AAdd { get; private set; }
+        public ushort RMult { get; private set; } = 1;
+        public ushort GMult { get; private set; } = 1;
+        public ushort BMult { get; private set; } = 1;
+        public ushort AMult { get; private set; } = 1;
+        public ushort RAdd { get; private set; }
+        public ushort GAdd { get; private set; }
+        public ushort BAdd { get; private set; }
+        public ushort AAdd { get; private set; }
 
         public bool HasAdd { get; private set; }
         public bool HasMult { get; private set; }
@@ -22,17 +22,17 @@ namespace CWAEmu.OFUCU.Flash.Records {
             cxform.nBits = reader.readUBits(4);
 
             if (cxform.HasMult) {
-                cxform.RMult = (byte)reader.readBits(cxform.nBits);
-                cxform.GMult = (byte)reader.readBits(cxform.nBits);
-                cxform.BMult = (byte)reader.readBits(cxform.nBits);
-                cxform.AMult = (byte)reader.readBits(cxform.nBits);
+                cxform.RMult = (ushort)reader.readBits(cxform.nBits);
+                cxform.GMult = (ushort)reader.readBits(cxform.nBits);
+                cxform.BMult = (ushort)reader.readBits(cxform.nBits);
+                cxform.AMult = (ushort)reader.readBits(cxform.nBits);
             }
 
             if (cxform.HasAdd) {
-                cxform.RAdd = (byte)reader.readBits(cxform.nBits);
-                cxform.GAdd = (byte)reader.readBits(cxform.nBits);
-                cxform.BAdd = (byte)reader.readBits(cxform.nBits);
-                cxform.AAdd = (byte)reader.readBits(cxform.nBits);
+                cxform.RAdd = (ushort)reader.readBits(cxform.nBits);
+                cxform.GAdd = (ushort)reader.readBits(cxform.nBits);
+                cxform.BAdd = (ushort)reader.readBits(cxform.nBits);
+                cxform.AAdd = (ushort)reader.readBits(cxform.nBits);
             }
 
             reader.endBitRead();
